SS-Incident-App-Prod/requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.5.1
mysqlclient==2.0.3
Brotli==1.0.9
cffi==1.14.6
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.11.2
django-import-export==2.5.0
django-phonenumber-field==5.1.0
et-xmlfile==1.1.0
filelock==3.0.12
fonttools==4.26.2
future==0.18.2
html5lib==1.1
isort==5.7.0
lazy-object-proxy==1.5.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
pbr==5.6.0
Pillow==8.3.1
pycparser==2.20
pydyf==0.0.3
pylint==2.7.2
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
six==1.15.0
sqlparse==0.4.1
stevedore==3.3.0
tablib==3.0.0
tinycss2==1.1.0
toml==0.10.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.8
